NavBar

+++ src/main/java/com/csa/samplefullstack/entity/employee.java
@@ -8,12 +8,12 @@ public class employee {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int employeeId;
+    private Long employeeId;
     private String name;
     private String department;
     private String location;
 
-    public employee(int employeeId, String name, String department, String location) {
+    public employee(Long employeeId, String name, String department, String location) {
         this.employeeId = employeeId;
         this.name = name;
         this.department = department;
@@ -24,11 +24,11 @@ public class employee {
 
     }
 
-    public int getEmployeeId() {
+    public Long getEmployeeId() {
         return employeeId;
     }
 
-    public void setEmployeeId(int employeeId) {
+    public void setEmployeeId(Long employeeId) {
         this.employeeId = employeeId;
     }
